% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{proxysearch}
\alias{proxysearch}
\title{Search for proxies}
\usage{
proxysearch(r2 = proxy_r2, d_input = d_mr)
}
\arguments{
\item{r2}{$r^2$ parameter for proxies}

\item{d_input}{Input data}
}
\description{
Function to search proxies for instrumental variables not found in the outcome summary statistics.
Not implemented yet.
}
